requirements.txt
python-telegram-bot>=20.0
python-dotenv
azure-cosmos
google-generativeai
flask
